♻️: removed useless views and repositories and renamed some folder
