+++ frontend/app_student/lib/api/week_schedule/repositories/week_schedule_repository.dart
@@ -10,7 +10,7 @@ class WeekScheduleRepository {
   Future<List<WeekScheduleModel>> getWeeksSchedule(className) {
     return apiService.getData('/api/timetable?class_param=$className', (item) {
       final entity = WeekScheduleEntity.fromJson(item);
-      return WeekScheduleModel(weekSchedule: entity);
+      return WeekScheduleModel.fromEntity(entity);
     });
   }
 }
